switched to integer units instead of floats

+++ hal-native/src/main/native/backends/bbb-usermode-sysfs/sysfs/pwm.h
@@ -8,8 +8,8 @@
 
 typedef struct {
     pwm_pin_t* pin;
-    unsigned period_ns;
-    unsigned duty_ns;
+    uint32_t period_ns;
+    uint32_t duty_ns;
 } pwm_t;
 
 
@@ -19,5 +19,5 @@ hal_error_t pwm_unexport(pwm_t* pwm);
 hal_error_t pwm_enable(pwm_t* pwm);
 hal_error_t pwm_disable(pwm_t* pwm);
 
-hal_error_t pwm_set_duty_cycle(pwm_t* pwm, float duty);
-hal_error_t pwm_set_frequency(pwm_t* pwm, float frequency);
+hal_error_t pwm_set_duty_cycle(pwm_t* pwm, uint32_t duty);
+hal_error_t pwm_set_frequency(pwm_t* pwm, uint32_t frequency);
